Created pde and ino files for live serial communication between processing and arduino (EDA and PPGIR signals measured using Emotibit)
The emotibit transmits data to the computer which is received using oscP5 in Processing. This is then transmitted to the arduino and received back.
The EDA signal is scaled by 10^6 times  to convert it into an int before transmitting to arduino  for now.
Graphing functionality has been removed in this code, but all data is printed on processing's console.
Note: The data being sent from arduino (Serial.println("") is also printed on the common serial monitor (processing's).